cw-cap: swap labels and icons to improve UX.
This change moves the master control icon to the left of the master connect
label to remove confusion about its use. Previously the icon was placed between
the master label and the scope label making it confusing as to where the icon
belongs to. For short "Master: X Scope: X Target: X" is replaced by
"X Master X Scope X Target".