Fix basic password check C file
As pointed out in issue 146 the delay_200_ms function does not work
when used on the XMEGA as it is an 8 bit microcontroller, where the
size of an int is 2 bytes, max number possible is 65635 with unsigned
so the previous end of loop condition of > 100000 would never be reached.
This is now fixed, and the timing works well with the SAFE-O-MATIC
ultra secure password protected system ;-).

Was tested on XMEGA12804 and STM32F3(32-bit).