Fix api target.protver not setting or getting
The SimpleSerial.py SimpleSerial class contained a local variable protver
that did not get or set the protocol version parameter in the back end.
This leads to confusion as it was often ignored when communicating with the
target. This is fixed. The protver is used to set the desired protocol and
the _protocol is the protocol version actually used. Setting protocol to
'auto' sets the _protver to the correct protocol version in the background.