Fix Python Console inability to end indented block
In the process of adding the ability to process multiple lines
of code when they are copied and pasted into the python console
I introduced a bug. The filtering of empty lines made it not possible
to end an indented block as it would remove the empty line. This is
now fixed.