Modify the chipwhisperer glitcher example code to be more easy to extend.
The ChipWhisperer glitcher example code is nice and easy to understand but on the other hand it is quite hard to modify to change or add more parameters. This rewrite of the example solves this problem by introducing a more generic parameter wrapper class at the cost of some complexity.